UPLOAD WORKING VERSION

+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/generators/GeneratorResource.java
@@ -11,9 +11,6 @@ import de.mephisto.vpin.server.util.Config;
 import de.mephisto.vpin.server.util.ImageUtil;
 import de.mephisto.vpin.server.util.RequestUtil;
 import edu.umd.cs.findbugs.annotations.NonNull;
-import org.apache.commons.fileupload.FileItem;
-import org.apache.commons.fileupload.disk.DiskFileItemFactory;
-import org.apache.commons.fileupload.servlet.ServletFileUpload;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.io.IOUtils;
 import org.slf4j.Logger;
@@ -25,12 +22,12 @@ import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpServletRequest;
 import java.awt.image.BufferedImage;
-import java.io.ByteArrayInputStream;
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.net.URLEncoder;
 import java.util.Arrays;
-import java.util.Iterator;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -93,60 +90,54 @@ public class GeneratorResource {
   }
 
   @PostMapping("/directb2supload")
-  public Boolean directb2supload(@RequestParam("file") MultipartFile file,
+  public Boolean directb2supload(@RequestParam(value = "file", required = false) MultipartFile file,
                                  @RequestParam("uploadType") String uploadType,
                                  @RequestParam("gameId") int gameId) {
-//    String name = file.getOriginalFilename().replaceAll("/", "").replaceAll("\\\\", "");
-//    File out = new File(systemService.getDirectB2SFolder(), name);
-//    if (uploadType.equals("generator")) {
-//      Game game = gameService.getGame(gameId);
-//      String directb2sFilename = FilenameUtils.getBaseName(game.getGameFileName()) + ".directb2s";
-//      out = new File(systemService.getVPXTablesFolder(), directb2sFilename);
-//    }
-//
-//    out.mkdirs();
-//    LOG.info("Uploading " + out.getAbsolutePath());
-//    return upload(file, out);
-    return false;
+    if (file == null) {
+      LOG.error("Upload request did not contain a file object.");
+      return false;
+    }
+
+    String name = file.getOriginalFilename().replaceAll("/", "").replaceAll("\\\\", "");
+    File out = new File(systemService.getDirectB2SFolder(), name);
+    if (!uploadType.equals("generator")) {
+      Game game = gameService.getGame(gameId);
+      String directb2sFilename = FilenameUtils.getBaseName(game.getGameFileName()) + ".directb2s";
+      out = new File(systemService.getVPXTablesFolder(), directb2sFilename);
+    }
+
+    out.mkdirs();
+    LOG.info("Uploading " + out.getAbsolutePath());
+    return upload(file, out);
   }
 
   @PostMapping(value = "/backgroundupload")
-  public Boolean upload(HttpServletRequest request) {
-    try {
-      boolean isMultipart = ServletFileUpload.isMultipartContent(request);
-      DiskFileItemFactory factory = new DiskFileItemFactory();
-      factory.setRepository(new File(System.getProperty("java.io.tmpdir")));
-      factory.setSizeThreshold(1024 * 1024 * 500);
-      factory.setFileCleaningTracker(null);
-
-      ServletFileUpload upload = new ServletFileUpload(factory);
-      List items = upload.parseRequest(request);
-      LOG.info("Upload has " + items.size() + " items.");
-
-      Iterator iter = items.iterator();
-      while (iter.hasNext()) {
-        FileItem item = (FileItem) iter.next();
-        if (!item.isFormField()) {
-          String name = item.getName();
-          File backgroundsFolder = new File(SystemService.RESOURCES, "backgrounds");
-          File out = new File(backgroundsFolder, name);
-          return upload(item, out);
-        }
-      }
-    } catch (Exception e) {
-      LOG.error("Upload failed: " + e.getMessage(), e);
+  public Boolean upload(@RequestPart(value = "file", required = false) MultipartFile file, HttpServletRequest request) throws IOException {
+    if (file == null) {
+      LOG.error("Upload request did not contain a file object.");
+      return false;
     }
-    return false;
+
+    String name = file.getOriginalFilename().replaceAll("/", "").replaceAll("\\\\", "");
+    File backgroundsFolder = new File(SystemService.RESOURCES, "backgrounds");
+    File out = new File(backgroundsFolder, name);
+    return upload(file, out);
   }
 
-  private Boolean upload(FileItem file, File target) {
+  private Boolean upload(MultipartFile file, File target) {
+    byte[] bytes = new byte[0];
     try {
-      byte[] bytes = IOUtils.toByteArray(file.getInputStream());
+      bytes = file.getBytes();
+      if (target.exists() && !target.delete()) {
+        throw new UnsupportedOperationException("Failed to delete existing target file " + target.getAbsolutePath());
+      }
+
       FileOutputStream fileOutputStream = new FileOutputStream(target);
       IOUtils.write(bytes, fileOutputStream);
       fileOutputStream.close();
+      LOG.info("Written uploaded file: " + target.getAbsolutePath() + ", byte size was " + bytes.length);
     } catch (Exception e) {
-      LOG.error("Failed to store asset: " + e.getMessage() + ", byte size was " + file.getSize(), e);
+      LOG.error("Failed to store asset: " + e.getMessage() + ", byte size was " + bytes.length, e);
     }
     return true;
   }
